created answerDao and moved dao logic and added comments
