+++ Trello_quora/quora-service/src/main/java/com/upgrad/quora/service/business/AnswerService.java
@@ -1,6 +1,7 @@
 package com.upgrad.quora.service.business;
 
 
+import com.upgrad.quora.service.dao.AnswerDao;
 import com.upgrad.quora.service.dao.UserDao;
 import com.upgrad.quora.service.entity.AnswerEntity;
 import com.upgrad.quora.service.entity.QuestionEntity;
@@ -20,6 +21,8 @@ import java.util.List;
 public class AnswerService {
     @Autowired
     private UserDao userDao;
+    @Autowired
+    private AnswerDao answerDao;
 
     @Transactional (propagation = Propagation.REQUIRED)
     public AnswerEntity createAnswer(AnswerEntity answerEntity, final String authorizationToken, final String questionId) throws AuthorizationFailedException, InvalidQuestionException {
@@ -33,7 +36,8 @@ public class AnswerService {
         if(userAuthTokenEntity.getLogoutAt() != null){
             throw new AuthorizationFailedException("ATHR-002","User is signed out.Sign in first to post an answer");
         }
-        QuestionEntity questionEntity = userDao.getQuestion(questionId);
+
+        QuestionEntity questionEntity = answerDao.getQuestion(questionId);
 
         //Validate if question is valid
         if (questionEntity == null){
@@ -41,7 +45,7 @@ public class AnswerService {
         }
         answerEntity.setUser(userAuthTokenEntity.getUser());
         answerEntity.setQuestion(questionEntity);
-        return userDao.createAnswer(answerEntity);
+        return answerDao.createAnswer(answerEntity);
     }
 
     @Transactional (propagation = Propagation.REQUIRED)
@@ -59,18 +63,18 @@ public class AnswerService {
 
         // If the answer with uuid which is to be edited does not exist in the database,
         // throw "AnswerNotFoundException" with the message code - 'ANS-001' and message - 'Entered answer uuid does not exist'.
-        if (userDao.getAnswer(answerId) == null){
+        if (answerDao.getAnswer(answerId) == null){
             throw new AnswerNotFoundException("ANS-001","Entered answer uuid does not exist");
         }
 
         // Only the answer owner can edit the answer. Therefore, if the user who is not the owner of the answer tries to edit the answer
         // throw "AuthorizationFailedException" with the message code - 'ATHR-003' and message - 'Only the answer owner can edit the answer'.
-        UserEntity answerOwner = userDao.getAnswer(answerId).getUser();
-        QuestionEntity questionEntity = userDao.getAnswer(answerId).getQuestion();
+        UserEntity answerOwner = answerDao.getAnswer(answerId).getUser();
+        QuestionEntity questionEntity = answerDao.getAnswer(answerId).getQuestion();
         if (answerOwner.getUuid().equals(userAuthTokenEntity.getUuid())){
             answerEntity.setUser(userAuthTokenEntity.getUser());
             answerEntity.setQuestion(questionEntity);
-            return userDao.editAnswer(answerEntity);
+            return answerDao.editAnswer(answerEntity);
         } else {
             throw new AuthorizationFailedException ("ATHR-003","Only the answer owner can edit the answer");
         }
@@ -89,16 +93,16 @@ public class AnswerService {
             throw new AuthorizationFailedException("ATHR-002", "User is signed out.Sign in first to delete an answer");
         }
         // Validate if the answer with uuid which is to be deleted does exist in the database,
-        if (userDao.getAnswer(answerId) == null) {
+        if (answerDao.getAnswer(answerId) == null) {
             throw new AnswerNotFoundException("ANS-001", "Entered answer uuid does not exist");
         }
 
         //Validate user is admin or owner of the answer or else throw an exception
         String role = userAuthToken.getUser().getRole();
-        UserEntity answerOwner = userDao.getAnswer(answerId).getUser();
+        UserEntity answerOwner = answerDao.getAnswer(answerId).getUser();
 
         if (answerOwner.getUuid().equals(userAuthToken.getUuid()) || role.equals("admin")) {
-            userDao.deleteAnswer(answerId);
+            answerDao.deleteAnswer(answerId);
         } else {
             throw new AuthorizationFailedException("ATHR-003", "Only the answer owner or admin can delete the answer");
         }
@@ -107,7 +111,7 @@ public class AnswerService {
     }
 
     public QuestionEntity getQuestion (final String questionId) {
-        return userDao.getQuestion(questionId);
+        return answerDao.getQuestion(questionId);
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
@@ -122,13 +126,13 @@ public class AnswerService {
         if(userAuthTokenEntity.getLogoutAt() != null){
             throw new AuthorizationFailedException("ATHR-002","User is signed out.Sign in first to get the answers");
         }
-        QuestionEntity questionEntity = userDao.getQuestion(questionId);
+        QuestionEntity questionEntity = answerDao.getQuestion(questionId);
         //Validate if question is valid
         if (questionEntity == null){
             throw new InvalidQuestionException("QUES-001","The question with entered uuid whose details are to be seen does not exist");
         }
-
-        List<AnswerEntity> allAnswers = userDao.getAllAnswers(questionId);
+        //If user is authorised and question is valid, return all answers along with answer id and question content
+        List<AnswerEntity> allAnswers = answerDao.getAllAnswers(questionId);
         return allAnswers;
 
     }
