+++ Trello_quora/quora-service/src/main/java/com/upgrad/quora/service/dao/UserDao.java
@@ -42,49 +42,5 @@ public class UserDao {
         UserEntity userEntity = getUser(userUuid);
         entityManager.remove(userEntity);
     }
-
-    public AnswerEntity createAnswer (AnswerEntity answerEntity){
-        entityManager.persist(answerEntity);
-        return answerEntity;
-    }
-
-    public AnswerEntity editAnswer (AnswerEntity answerEntity){
-        entityManager.merge(answerEntity);
-        return answerEntity;
-    }
-    public QuestionEntity getQuestion(final String questionUuid){
-        try {
-            return entityManager.createNamedQuery("questionByUuid", QuestionEntity.class)
-                    .setParameter("uuid", questionUuid)
-                    .getSingleResult();
-        }catch (NoResultException nre){
-            return null;
-        }
-    }
-
-    public AnswerEntity getAnswer(final String answerUuid){
-        try {
-            return entityManager.createNamedQuery("answerByUuid", AnswerEntity.class)
-                    .setParameter("uuid", answerUuid)
-                    .getSingleResult();
-        }catch (NoResultException nre){
-            return null;
-        }
-    }
-    public void deleteAnswer(final String AnswerId){
-        AnswerEntity answerEntity = getAnswer(AnswerId);
-        entityManager.remove(answerEntity);
-    }
-
-    public List<AnswerEntity> getAllAnswers (final String questionId){
-        QuestionEntity questionEntity = getQuestion(questionId);
-        try {
-            return entityManager.createNamedQuery("answersByQuesID", AnswerEntity.class)
-                    .setParameter("question_id", questionEntity)
-                    .getResultList();
-        }catch (NoResultException nre){
-            return null;
-        }
-    }
 }
